refactoring ClientController
