+++ src/main/java/ru/geekbrains/pocket/messenger/client/view/ChatViewController.java
@@ -2,6 +2,9 @@ package ru.geekbrains.pocket.messenger.client.view;
 
 import ru.geekbrains.pocket.messenger.client.Main;
 import ru.geekbrains.pocket.messenger.client.controller.ClientController;
+import ru.geekbrains.pocket.messenger.client.controller.ContactController;
+import ru.geekbrains.pocket.messenger.client.controller.GroupController;
+import ru.geekbrains.pocket.messenger.client.controller.MessageController;
 import ru.geekbrains.pocket.messenger.client.utils.Common;
 import ru.geekbrains.pocket.messenger.client.utils.CustomTextArea;
 import ru.geekbrains.pocket.messenger.client.view.customFX.*;
@@ -53,7 +56,18 @@ import java.util.ResourceBundle;
 public class ChatViewController implements Initializable {
 
     private static ChatViewController instance;
+    private ClientController clientController;
+    private ContactController contactController;
+    private MessageController messageController;
+    private GroupController groupController;
 
+    private WebEngine webEngine;
+    private ObservableList<CFXListElement> contactsObservList;
+    private String backgroundImage;
+    private Document DOMdocument;
+    private String tsOld;
+    private int idDivMsg;
+    private int idMsg;
 
     @FXML
     private BorderPane borderPaneMain;
@@ -137,22 +151,6 @@ public class ChatViewController implements Initializable {
 
     @FXML
     private JFXTabPane tabPane;
-    //
-    private WebEngine webEngine;
-
-    private ObservableList<CFXListElement> contactsObservList;
-
-    private ClientController clientController;
-
-    private String backgroundImage;
-
-    private Document DOMdocument;
-
-    private String tsOld;
-
-    private int idDivMsg;
-
-    private int idMsg;
 
     @FXML
     private  JFXButton btnContactSearchCancel;
@@ -206,8 +204,11 @@ public class ChatViewController implements Initializable {
         initWebView();
 
         clientController = ClientController.getInstance();
-        clientController.setChatViewController(this);
-        contactsObservList = FXCollections.observableList(clientController.getContactListOfCards());
+        contactController = ContactController.getInstance();
+        messageController = MessageController.getInstance();
+        groupController = GroupController.getInstance();
+        contactController.setChatViewController(this);
+        contactsObservList = FXCollections.observableList(contactController.getContactListOfCards());
         // при пустом списке контактов открыть вкладку контакты //todo перепилить на список чатов?
         if (contactsObservList.isEmpty()) contacts.getTabPane().getSelectionModel().select(contacts);
         contactListView.setExpanded(true);
@@ -223,7 +224,7 @@ public class ChatViewController implements Initializable {
                 String text = messageField.getText().trim();
                 if (!text.isEmpty()) {
 //                    messageField.appendText(System.lineSeparator());
-                    clientController.sendMessage(messageField.getText());
+                    messageController.sendMessage(clientController.getMyUser(), contactController.getReciever(), messageField.getText());
                     messageField.clear();
                     messageField.requestFocus();
                 }
@@ -555,7 +556,7 @@ public class ChatViewController implements Initializable {
     @FXML
     private void handleSendMessage() {
         if (!messageField.getText().isEmpty()) {
-            clientController.sendMessage(messageField.getText());
+            messageController.sendMessage(clientController.getMyUser(), contactController.getReciever(), messageField.getText());
             messageField.clear();
             messageField.requestFocus();
         }
@@ -566,7 +567,7 @@ public class ChatViewController implements Initializable {
         String receiver = contactListView.getSelectionModel().getSelectedItem().getUser().getUid();
         if (event.getClickCount() == 1) {
             //showAlert("Сообщения будут отправляться контакту " + receiver, Alert.AlertType.INFORMATION);
-            clientController.setReceiver(receiver);
+            contactController.setReceiver(receiver);
             messageField.requestFocus();
             messageField.selectEnd();
         } else if (event.getClickCount() == 2) {
@@ -582,9 +583,9 @@ public class ChatViewController implements Initializable {
     private void handleFindedClientChoice(MouseEvent event) {
         String receiver = searchListView.getSelectionModel().getSelectedItem().getUser().getUid();
         if (event.getClickCount() == 1) {
-            if (clientController.hasReceiver(receiver)) {
+            if (contactController.hasReceiver(receiver)) {
                 btnContactSearchInvite.setVisible(false);
-                clientController.setReceiver(receiver);
+                contactController.setReceiver(receiver);
                 messageField.requestFocus();
                 messageField.selectEnd();
             } else {
@@ -594,7 +595,7 @@ public class ChatViewController implements Initializable {
         } else if (event.getClickCount() == 2) {
             othersProfile.setUser(
                     searchListView.getSelectionModel().getSelectedItem().getUser());
-            othersProfile.setIfFriendly(clientController.hasReceiver(receiver));
+            othersProfile.setIfFriendly(contactController.hasReceiver(receiver));
             PaneProvider.setProfileScrollPane(otherProfileScrollPane);
             paneProvidersProfScrollPaneVisChange(true);
         }
@@ -696,7 +697,7 @@ public class ChatViewController implements Initializable {
 
         for (File fs : f.listFiles()) {
             img += fs.toURI();
-            clientController.sendMessage(img);
+            messageController.sendMessage(clientController.getMyUser(), contactController.getReciever(), img);
             webEngine.executeScript("document.getElementById(\"" + idMsg + "\").innerHTML = '" + "<img src = \"" + img + "\" width=\"30\" alt=\"lorem\"/>" +"'");
             setIdMsg(idMsg++);
             webEngine.executeScript("document.getElementById(\"" + (getIdMsg()) + "\").innerHTML = '" + "<img src = \"" + img + "\" width=\"30\" alt=\"lorem\"/>" +"'");
@@ -837,7 +838,7 @@ public class ChatViewController implements Initializable {
 
     @FXML
     public void handleGroupCreateButton(){
-        clientController.addGroup(creategroupName.getText());
+        groupController.addGroup(creategroupName.getText(), clientController.getToken());
     }
 
     @FXML
@@ -857,7 +858,7 @@ public class ChatViewController implements Initializable {
             });
             // todo: поиск на сервере от 2х символов, убрать/расширить ограничение?
             if (tfSearchInput.getText().length()>=6) {
-                CFXListElement searchFromServer = clientController.findContact(tfSearchInput.getText());
+                CFXListElement searchFromServer = contactController.findContact(tfSearchInput.getText(), clientController.getToken());
                 //todo: статус пользователей (онлайн/офлайн) - будет приходить с сервера или запрашивать на каждого?
 //                if (searchFromServer != null) {
 //                    searchFromServer.removeAll(searchObsList);
@@ -883,7 +884,7 @@ public class ChatViewController implements Initializable {
 
     @FXML
     private void contactSearchBtnInviteClicked() {
-        clientController.addContact(searchListView.getSelectionModel().getSelectedItem().getUser());
+        contactController.addContact(searchListView.getSelectionModel().getSelectedItem().getUser());
         contactSearchBtnCancelClicked();
     }
