+++ src/main/java/ru/geekbrains/pocket/messenger/client/view/customFX/CFXOtherProfile.java
@@ -1,6 +1,7 @@
 package ru.geekbrains.pocket.messenger.client.view.customFX;
 
 import ru.geekbrains.pocket.messenger.client.controller.ClientController;
+import ru.geekbrains.pocket.messenger.client.controller.ContactController;
 import ru.geekbrains.pocket.messenger.client.view.PaneProvider;
 import com.jfoenix.controls.JFXButton;
 import com.jfoenix.controls.JFXTextArea;
@@ -16,12 +17,10 @@ import java.io.IOException;
 public class CFXOtherProfile extends AnchorPane {
 
     private boolean isFriend = false;
-
     private User user;
 
     @FXML
     private AnchorPane otherProfilePane;
-
     @FXML
     private JFXButton btnCloseRight;
     @FXML
@@ -40,16 +39,12 @@ public class CFXOtherProfile extends AnchorPane {
     private Label labelName;
     @FXML
     private JFXTextArea textareaInfo;
-
     @FXML
     private JFXButton writeMsgBtn;
-    
     @FXML
     private JFXButton notificationsBtn;
-
     @FXML
     private JFXButton clearMsgsBtn;
-
     @FXML
     private JFXButton removeUserBtn;
 
@@ -121,23 +116,23 @@ public class CFXOtherProfile extends AnchorPane {
     }
 
     private void writeMsg() {
-        ClientController.getInstance().setReceiver(user);
+        ContactController.getInstance().setReceiver(user);
     }
 
     private void notificationsConf() {
     }
 
     private void clearMsgs() {
-        ClientController.getInstance().clearMessagesWithUser(user); // todo подтверждение?
+        ContactController.getInstance().clearMessagesWithUser(user); // todo подтверждение?
     }
 
     private void removeUser() {
         closeButtonPressed();
-        ClientController.getInstance().removeContact(user); // todo подтверждение?
+        ContactController.getInstance().removeContact(user); // todo подтверждение?
     }
 
     private void inviteContact() {
         closeButtonPressed();
-        ClientController.getInstance().addContact(user);
+        ContactController.getInstance().addContact(user);
     }
 }
